requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

azure-storage-blob
google-analytics-data
google-cloud-core
google-auth
azure-keyvault-secrets
azure-identity
pandas
requests
cygrpc